+++ stdlib/lib.core.d.ts
@@ -679,6 +679,9 @@ interface ProcessWritableStream {
     readonly writable: boolean;
     addListener(eventName: string, listener: (...args: any[]) => void, ...ignored: any[]): void;
     cork(...ignored: any[]): void;
+    end(callback?: () => void): void;
+    end(chunk: string | Buffer, callback?: () => void): void;
+    end(chunk: string | Buffer, encoding?: string, callback?: () => void): void;
     off(eventName: string, listener: (...args: any[]) => void, ...ignored: any[]): void;
     on(eventName: string, listener: (...args: any[]) => void, ...ignored: any[]): void;
     once(eventName: string, listener: (...args: any[]) => void, ...ignored: any[]): void;